LEER DATOS DE UNA ESTACION METEOROLOGICA

Descarga (si es necesario) y lee el archivo CSV de una estacion meteorologica del conjunto de estaciones disponibles.

Usage

leer_datos_estacion(id_estacion, ruta_archivo)

Arguments

id_estacion

Cadena de texto que identifica la estacion a leer.

ruta_archivo

Ruta local donde se almacenara o desde donde se leera el archivo CSV.

Value

Un objeto de clase data.frame (tibble) con los registros meteorologicos de la estacion seleccionada. Cada fila representa una observacion diaria.

Details

La funcion verifica si el archivo correspondiente a la estacion ya esta descargado en el sistema local. Si no existe, lo descarga desde el repositorio en linea de referencia. Luego, lee los datos y devuelve un data frame con las observaciones meteorologicas.

Examples

# Descargar y leer los datos de la estacion NH0472
datos <- leer_datos_estacion("NH0472", "datos/NH0472.csv")
#> El archivo no esta descargado, se procede a descargarlo...
#> Lectura completada correctamente para la estacion NH0472.

# Leer el archivo si ya esta descargado
datos_local <- leer_datos_estacion("NH0472", "datos/NH0472.csv")
#> El archivo ya esta descargado, se procede a leerlo...
#> Lectura completada correctamente para la estacion NH0472.