TABLA RESUMEN DE TEMPERATURA

Calcula estadisticas descriptivas de la variable temperatura_abrigo_150cm para una o varias estaciones meteorologicas y devuelve una tabla resumen con los resultados.

Usage

tabla_resumen_temperatura(...)

Arguments

...

Uno o mas data frames con registros meteorologicos que contengan la columna temperatura_abrigo_150cm y el identificador de estacion id.

Value

Un objeto de clase data.frame con las siguientes columnas:

estacion: identificador unico de la estacion (id).

media: temperatura media (°C).

minimo: temperatura minima registrada (°C).

maximo: temperatura maxima registrada (°C).

desviacion: desviacion estandar de la temperatura (°C).

n_observaciones: cantidad de registros validos considerados en el calculo.

Cada fila representa el resumen estadistico de una estacion meteorologica.

Details

La funcion valida que todos los argumentos recibidos sean data frames y luego calcula para cada uno un conjunto de medidas descriptivas de (temperatura_abrigo_150cm). Devuelve una tabla combinada con un resumen por estacion.

Examples

# Calcular resumen de una sola estacion
data(NH0472)
resumen_NH0472 <- tabla_resumen_temperatura(NH0472)
#>   estacion    media minimo maximo desviacion n_observaciones
#> 1   NH0472 18.02932   0.55   34.8   5.937767           20359
#> 
#> Resumen generado correctamente para 1 estacion(es).

# Calcular resumen combinado para varias estaciones
data(NH0472)
data(NH0910)
data(NH0046)
resumen_multiple <- tabla_resumen_temperatura(NH0472, NH0910, NH0046)
#>   estacion    media minimo maximo desviacion n_observaciones
#> 1   NH0472 18.02932   0.55   34.8   5.937767           20359
#> 
#>   estacion    media minimo maximo desviacion n_observaciones
#> 1   NH0910 17.65387      0  35.25   6.117994           15506
#> 
#>   estacion    media minimo maximo desviacion n_observaciones
#> 1   NH0046 17.28705  -0.25   34.6   5.895503           16782
#> 
#> Resumen generado correctamente para 3 estacion(es).